cl hw2 bug fixed

+++ fall_2014/cl/hw2/hw2/main.cpp
@@ -82,7 +82,7 @@ public:
 
         // read result
         std::vector<float> output(input.size());
-        queue.enqueueReadBuffer(dev_output, CL_TRUE, 0, sizeof(int) * input.size(), &output[0]);
+        queue.enqueueReadBuffer(dev_output, CL_TRUE, 0, sizeof(float) * input.size(), &output[0]);
 
         print_res(os, output);
     }
